Add files via upload
I have integrated the code that sets up the grid_map and the beacon_map. Currently they are two separate grid_maps. I think this should be just fine. Sam, the drone would be moving in the grid_map, but checking  beacon_map at the same locations for a signal